+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.0.2
 
-- Implmented directive generators (`export`, `import`, `library`, `part`, `part of`)
-- In development script generators (`library`, `library part`)
+- Implemented directive generators (`export`, `import`, `library`, `part`, `part of`)
+- Implemented script generators (`library`, `library part`)
 
 ## 0.0.1